added counter and unabled sending empty todos
