+++ code/src/components/CheckBox.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 
-import styled from 'styled-components'
+import { Icon, CheckboxContainer, HiddenCheckbox, StyledCheckbox } from "../styling/styleCheckBox"
 
 export const CheckBox = ({ isChecked, onChangeHandler }) => {
   const [checked, setChecked] = useState(isChecked)
@@ -19,7 +19,7 @@ export const CheckBox = ({ isChecked, onChangeHandler }) => {
         </Icon>
       </StyledCheckbox>
     </CheckboxContainer>
-  );
+  )
 
   return (
     <div>
@@ -33,38 +33,3 @@ export const CheckBox = ({ isChecked, onChangeHandler }) => {
   );
 };
 
-const Icon = styled.svg`
-  fill: none;
-  stroke: white;
-  stroke-width: 2px;
- `
-
-const HiddenCheckbox = styled.input.attrs({ type: 'checkbox' })` 
-  border: 0;
-  clip:rect();
-  clip-path: inset(50%);
-  height: 1px;
-  margin: -1px;
-  overflow: hidden;
-  padding: 0;
-  position: absolute;
-  white-space: nowrap;
-  width: 1px;
-`
-
-const StyledCheckbox = styled.div`
-  display: inline-block;
-  width: 25px;
-  height: 25px;
-  margin-right: 20px;
-  border: 1px solid grey;
-  background: ${(props) => (props.checked) ? '#3CB371' : '#fdc1c5'};
-  transition: all 150ms; 
-  ${Icon} {
-      visibility: ${props => (props.checked ? 'visible' : 'hidden')}
-    }
-`
-const CheckboxContainer = styled.div`
-  display: inline-block;
-  vertical-align: middle;
-`
