+++ code/src/components/Header.js
@@ -14,7 +14,7 @@ export const Header = () => {
 }
 
 const HeaderWrapper = styled.section`
-  width:600px;
+  width:100vw;
   height: 100px;
   background: #f1f1f1;
   display: flex;
@@ -23,15 +23,25 @@ const HeaderWrapper = styled.section`
   align-items: center;
   padding: 10px;
   opacity:0.5;
+
+  @media (max-width: 413px) {
+    flex-direction: column;   
+  }
 `
 
 const Text = styled.p`
   font-size: 40px;
   font-weight: bold;
   font-family: 'Acme', sans-serif;
+  @media (max-width: 413px) {
+    margin: 0;
+  }
 `
 const Date = styled.p`
   font-size: 25px;
   margin-left: 25px;
   font-family: 'Roboto Mono', monospace;
+  @media (max-width: 413px) {
+    margin: 0;
+  }
 `
