+++ code/src/styling/styleAddTask.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components'
+
+export const Form = styled.form`
+  display: flex;
+  justify-content: center;
+  padding-top: 20px;
+`
+export const InputArea = styled.input`
+  height: 30px;
+`
+export const AddButton = styled.button`
+  background-color: #e6e6e6; 
+  padding: 5px;
+  border-radius: 50%;
+  height: 30px;
+  width: 30px;
+  align-items: center;
+  justify-content: center;
+  margin-left: 10px;
+`
